tools/evil-winrm/requirements.txt
mcp>=1.0.0
pywinrm>=0.4.3
